src/models/volatility_model.py
"""
Volatility modeling module implementing various volatility forecasting models.
"""

import numpy as np
import pandas as pd
from arch import arch_model
from typing import Optional, Tuple, Dict
from sklearn.base import BaseEstimator
from scipy.stats import norm

class VolatilityModel(BaseEstimator):
    
    def __init__(self):
        self.model = None
        self.fitted_model = None
        
    def fit(self, returns: pd.Series) -> 'VolatilityModel':
        """
        Fit the volatility model to the data.
        
        Args:
            returns: Series of returns
            
        Returns:
            self
        """
        raise NotImplementedError
        
    def predict(self, returns: pd.Series) -> pd.Series:
        """
        Predict volatility for the given returns.
        
        Args:
            returns: Series of returns
            
        Returns:
            Series of predicted volatilities
        """
        raise NotImplementedError

class GARCHModel(VolatilityModel):
    
    def __init__(
        self,
        p: int = 1,
        q: int = 1,
        mean: str = 'Zero',
        vol: str = 'GARCH',
        dist: str = 'normal'
    ):
        """
        Initialize GARCH model.
        
        Args:
            p: ARCH order
            q: GARCH order
            mean: Mean model specification
            vol: Volatility model specification
            dist: Error distribution
        """
        super().__init__()
        self.p = p
        self.q = q
        self.mean = mean
        self.vol = vol
        self.dist = dist
        
    def fit(self, returns: pd.Series) -> 'GARCHModel':
        """
        Fit GARCH model to the data.
        
        Args:
            returns: Series of returns
            
        Returns:
            self
        """
        self.model = arch_model(
            returns,
            p=self.p,
            q=self.q,
            mean=self.mean,
            vol=self.vol,
            dist=self.dist
        )
        self.fitted_model = self.model.fit(disp='off')
        return self
        
    def predict(
        self,
        returns: pd.Series,
        horizon: int = 1
    ) -> Tuple[pd.Series, Dict]:
        """
        Predict volatility and return model parameters.
        
        Args:
            returns: Series of returns
            horizon: Forecast horizon
            
        Returns:
            Tuple of (predicted volatility, model parameters)
        """
        if self.fitted_model is None:
            raise ValueError("Model must be fitted before prediction")
            
        forecast = self.fitted_model.forecast(horizon=horizon)
        params = self.fitted_model.params.to_dict()
        
        return forecast.variance.iloc[-1], params
    
    def calculate_var(
        self,
        returns: pd.Series,
        confidence_level: float = 0.95
    ) -> float:
        """
        Calculate VaR.
        
        Args:
            returns: Series of returns
            confidence_level: VaR confidence level
            
        Returns:
            VaR estimate
        """
        if self.fitted_model is None:
            raise ValueError("Model must be fitted before VaR calculation")
            
        forecast = self.fitted_model.forecast(horizon=1)
        vol = np.sqrt(forecast.variance.iloc[-1])
        
        if self.dist == 'normal':
            z_score = norm.ppf(1 - confidence_level)
        else:
            # Handle other distributions
            raise NotImplementedError
            
        return -z_score * vol 