requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
alpaca-trade-api>=2.3.0
fredapi>=0.5.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
duckdb>=0.7.0

# ETL & Orchestration
apache-airflow>=2.5.0
prefect>=2.0.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.7.0
lightgbm>=3.3.0
tensorflow>=2.10.0
torch>=1.13.0
arch>=5.0.0
prophet>=1.1.0

# MLOps
mlflow>=2.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
dvc>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.13.0
altair>=4.2.0
streamlit>=1.22.0
dash>=2.9.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
jupyter>=1.0.0

# Documentation
sphinx>=6.1.0
sphinx-rtd-theme>=1.2.0 